修复帐号类型参数错误

+++ src/Senparc.Weixin.Open/Senparc.Weixin.Open/ComponentAPIs/ComponentApi.cs
@@ -113,7 +113,7 @@ namespace Senparc.Weixin.Open.ComponentAPIs
                     componentAppId.AsUrlData(), preAuthCode.AsUrlData(), redirectUrl.AsUrlData());
 
             if (authType != LoginAuthType.默认)
-                url = string.Format("{0}&auth_type={1}", url, authType);
+                url = string.Format("{0}&auth_type={1}", url, (int) authType);
 
             if (!string.IsNullOrEmpty(bizAppId))
                 url = string.Format("{0}&biz_appid={1}", url, bizAppId);
